LPS-81805 complete move to api module
