+++ modules/apps/portal-security-audit/portal-security-audit-api/bnd.bnd
@@ -3,4 +3,5 @@ Bundle-SymbolicName: com.liferay.portal.security.audit.api
 Bundle-Version: 3.0.0
 Export-Package:\
 	com.liferay.portal.security.audit,\
-	com.liferay.portal.security.audit.configuration
+	com.liferay.portal.security.audit.configuration,\
+	com.liferay.portal.security.audit.formatter
