+++ modules/apps/portal-security-audit/portal-security-audit-router/src/main/java/com/liferay/portal/security/audit/router/internal/CSVLogMessageFormatter.java
@@ -19,7 +19,7 @@ import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.audit.AuditMessage;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.security.audit.router.LogMessageFormatter;
+import com.liferay.portal.security.audit.formatter.LogMessageFormatter;
 import com.liferay.portal.security.audit.router.configuration.CSVLogMessageFormatterConfiguration;
 
 import java.util.Map;
