+++ modules/apps/portal-security-audit/portal-security-audit-router/src/main/java/com/liferay/portal/security/audit/router/internal/JSONLogMessageFormatter.java
@@ -16,7 +16,7 @@ package com.liferay.portal.security.audit.router.internal;
 
 import com.liferay.portal.kernel.audit.AuditMessage;
 import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.security.audit.router.LogMessageFormatter;
+import com.liferay.portal.security.audit.formatter.LogMessageFormatter;
 
 import org.osgi.service.component.annotations.Component;
